requirements.txt
pandas== 2.3.2
streamlit==1.49.0
plotly.express==0.4.1
python-dateutil==2.9.0.post0